+++ gen/gen.py
@@ -481,6 +481,8 @@ void fsh(const FReg& rs2, const Reg& rs1, int32_t imm12 = 0) { opStoreFP(0x1027,
 # misc
 print('''
 void nop() { if (supportRVC_) { append2B(0x0001); return; } addi(x0, x0, 0); }
+#ifdef XBYAK_RISCV_LI_OLD
+// imm is a 32-bit value (sign-extended on RV64)
 void li(const Reg& rd, uint32_t imm)
 {
 	if (imm && (imm & local::mask(12)) == 0) { // lower 12 bits of imm are zero
@@ -499,6 +501,20 @@ void li(const Reg& rd, uint32_t imm)
 		addiw(rd, rd, L);
 	}
 }
+#else
+// load a 64-bit immediate in the same way as LLVM
+// Zba/Zbb/Zbs instructions are used if supportBext(true)
+// RV32 : imm must be a 32-bit value (zero- or sign-extended)
+void li(const Reg& rd, uint64_t imm)
+{
+	if (isRV32_) {
+		const uint64_t hi = imm >> 32;
+		if (hi != 0 && hi != local::mask(32)) XBYAK_RISCV_THROW(ERR_IMM_IS_TOO_BIG)
+		imm = local::sext32(imm);
+	}
+	li_inner(rd, imm);
+}
+#endif
 void mv(const Reg& rd, const Reg& rs) { addi(rd, rs, 0); }
 void not_(const Reg& rd, const Reg& rs) { xori(rd, rs, -1); }
 void neg(const Reg& rd, const Reg& rs) { sub(rd, x0, rs); }